more gameboy opcode fixes.

+++ src/operations.rs
@@ -195,7 +195,7 @@ pub fn execute_operation(opcode: u8, cpu: &mut Cpu, mmu: &mut Mmu) -> () {
         }
         0x20 => {
             if !cpu.get_f_zero() {
-                let offset = mmu.read_byte(cpu.pc + 1);
+                let offset = mmu.read_byte(cpu.pc + 1) as i8;
                 cpu.pc = ((cpu.pc as u32 as i32) + (offset as i32)) as u16;
 
                 cpu.cycles += 1;
@@ -252,7 +252,7 @@ pub fn execute_operation(opcode: u8, cpu: &mut Cpu, mmu: &mut Mmu) -> () {
         }
         0x28 => {
             if cpu.get_f_zero() {
-                let offset = mmu.read_byte(cpu.pc + 1);
+                let offset = mmu.read_byte(cpu.pc + 1) as i8;
                 cpu.pc = ((cpu.pc as u32 as i32) + (offset as i32)) as u16;
 
                 cpu.cycles += 1;
@@ -308,7 +308,7 @@ pub fn execute_operation(opcode: u8, cpu: &mut Cpu, mmu: &mut Mmu) -> () {
         }
         0x30 => {
             if !cpu.get_f_carry() {
-                let offset = mmu.read_byte(cpu.pc + 1);
+                let offset = mmu.read_byte(cpu.pc + 1) as i8;
                 cpu.pc = ((cpu.pc as u32 as i32) + (offset as i32)) as u16;
 
                 cpu.cycles += 1;
@@ -371,7 +371,7 @@ pub fn execute_operation(opcode: u8, cpu: &mut Cpu, mmu: &mut Mmu) -> () {
         }
         0x38 => {
             if cpu.get_f_carry() {
-                let offset = mmu.read_byte(cpu.pc + 1);
+                let offset = mmu.read_byte(cpu.pc + 1) as i8;
                 cpu.pc = ((cpu.pc as u32 as i32) + (offset as i32)) as u16;
 
                 cpu.cycles += 1;
